+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/toolbar/index.tsx
@@ -20,6 +20,7 @@ import * as MediaClientModule from '@atlaskit/media-client';
 import { FileState, MediaClient } from '@atlaskit/media-client';
 import {
   asMockReturnValue,
+  expectFunctionToHaveBeenCalledWith,
   fakeMediaClient,
   getDefaultMediaClientConfig,
 } from '@atlaskit/media-test-helpers';
@@ -27,10 +28,7 @@ import { shallow } from 'enzyme';
 import { ReactElement } from 'react';
 import { IntlProvider } from 'react-intl';
 import commonMessages from '../../../../../messages';
-import {
-  FloatingToolbarButton,
-  FloatingToolbarCustom,
-} from '../../../../../plugins/floating-toolbar/types';
+import { FloatingToolbarCustom } from '../../../../../plugins/floating-toolbar/types';
 import Button from '../../../../../plugins/floating-toolbar/ui/Button';
 import { MediaOptions } from '../../../../../plugins/media';
 import {
@@ -42,7 +40,6 @@ import {
   AnnotationToolbar,
   messages as annotateMessages,
 } from '../../../../../plugins/media/toolbar/annotation';
-import { Command } from '../../../../../types';
 import { setNodeSelection } from '../../../../../utils';
 import {
   getFreshMediaProvider,
@@ -50,6 +47,10 @@ import {
   testCollectionName,
 } from '../_utils';
 import { ProviderFactory } from '@atlaskit/editor-common';
+import {
+  getToolbarItems,
+  findToolbarBtn,
+} from '../../floating-toolbar/_helpers';
 
 describe('media', () => {
   const createEditor = createEditorFactory<MediaPluginState>();
@@ -114,8 +115,9 @@ describe('media', () => {
 
       const toolbar = floatingToolbar(editorView.state, intl);
       expect(toolbar).toBeDefined();
-      const removeButton = toolbar!.items.find(
-        item => item.type === 'button' && item.title === removeTitle,
+      const removeButton = findToolbarBtn(
+        getToolbarItems(toolbar!, editorView),
+        removeTitle,
       );
 
       expect(removeButton).toBeDefined();
@@ -129,7 +131,7 @@ describe('media', () => {
       const { editorView } = editor(docWithMediaSingle);
 
       const toolbar = floatingToolbar(editorView.state, intl, {
-        appearance: 'full-page',
+        allowAdvancedToolBarOptions: true,
       });
       expect(toolbar).toBeDefined();
       expect(toolbar!.items.length).toEqual(11);
@@ -140,7 +142,7 @@ describe('media', () => {
 
       const toolbar = floatingToolbar(editorView.state, intl, {
         allowResizing: true,
-        appearance: 'full-page',
+        allowAdvancedToolBarOptions: true,
       });
       expect(toolbar).toBeDefined();
       expect(toolbar!.items.length).toEqual(8);
@@ -152,11 +154,13 @@ describe('media', () => {
       const toolbar = floatingToolbar(editorView.state, intl, {
         allowResizing: true,
         allowAnnotation: true,
-        appearance: 'full-page',
+        allowAdvancedToolBarOptions: true,
       });
       expect(toolbar).toBeDefined();
       expect(toolbar!.items.length).toEqual(9);
-      const item = toolbar!.items.find(cmd => cmd.type === 'custom');
+      const item = getToolbarItems(toolbar!, editorView).find(
+        cmd => cmd.type === 'custom',
+      );
       expect(item).toBeDefined();
     });
 
@@ -165,7 +169,7 @@ describe('media', () => {
 
       const toolbar = floatingToolbar(editorView.state, intl, {
         allowResizing: true,
-        appearance: 'comment',
+        allowAdvancedToolBarOptions: false,
       });
       expect(toolbar).toBeDefined();
       expect(toolbar!.items.length).toEqual(1);
@@ -183,7 +187,7 @@ describe('media', () => {
 
       const toolbar = floatingToolbar(editorView.state, intl, {
         allowResizing: true,
-        appearance: 'full-page',
+        allowAdvancedToolBarOptions: true,
       });
       expect(toolbar).toBeDefined();
       expect(toolbar!.items.length).toEqual(1);
@@ -201,7 +205,7 @@ describe('media', () => {
 
       const toolbar = floatingToolbar(editorView.state, intl, {
         allowResizing: true,
-        appearance: 'full-page',
+        allowAdvancedToolBarOptions: true,
       });
       expect(toolbar).toBeDefined();
       expect(toolbar!.items.length).toEqual(1);
@@ -212,7 +216,7 @@ describe('media', () => {
 
       const toolbar = floatingToolbar(editorView.state, intl, {
         allowResizing: true,
-        appearance: 'full-page',
+        allowAdvancedToolBarOptions: true,
       });
       expect(toolbar).toBeDefined();
       expect(toolbar!.items.length).toEqual(1);
@@ -225,7 +229,7 @@ describe('media', () => {
 
       const toolbar = floatingToolbar(editorView.state, intl, {
         allowResizing: true,
-        appearance: 'full-page',
+        allowAdvancedToolBarOptions: true,
       });
       expect(toolbar).toBeDefined();
       expect(toolbar!.items.length).toEqual(1);
@@ -236,9 +240,10 @@ describe('media', () => {
       setNodeSelection(editorView, 0);
 
       const toolbar = floatingToolbar(editorView.state, intl);
-      const removeButton = toolbar!.items.find(
-        item => item.type === 'button' && item.title === removeTitle,
-      ) as FloatingToolbarButton<Command>;
+      const removeButton = findToolbarBtn(
+        getToolbarItems(toolbar!, editorView),
+        removeTitle,
+      );
 
       removeButton.onClick(editorView.state, editorView.dispatch);
       expect(editorView.state.doc).toEqualDocument(doc(p()));
@@ -252,11 +257,13 @@ describe('media', () => {
 
       const toolbar = floatingToolbar(editorView.state, intl, {
         allowResizing: true,
-        appearance: 'full-page',
+        allowAdvancedToolBarOptions: true,
       });
-      const button = toolbar!.items.find(
-        item => item.type === 'button' && item.title === alignLeftTitle,
-      ) as FloatingToolbarButton<Command>;
+
+      const button = findToolbarBtn(
+        getToolbarItems(toolbar!, editorView),
+        alignLeftTitle,
+      );
 
       button.onClick(editorView.state, editorView.dispatch);
       expect(editorView.state.doc).toEqualDocument(
@@ -294,6 +301,22 @@ describe('media', () => {
         jest.resetAllMocks();
       });
 
+      it('should call getCurrentState for current state', () => {
+        shallow(
+          <AnnotationToolbar
+            viewMediaClientConfig={mockMediaClient.config}
+            id="1234"
+            collection="some-collection"
+            intl={intl}
+          />,
+        );
+
+        expectFunctionToHaveBeenCalledWith(
+          mockMediaClient.file.getCurrentState,
+          ['1234', { collectionName: 'some-collection' }],
+        );
+      });
+
       it('has an AnnotationToolbar custom toolbar element', async () => {
         const { editorView, pluginState } = editor(docWithMediaSingle);
         await pluginState.setMediaProvider(getFreshMediaProvider());
@@ -303,12 +326,13 @@ describe('media', () => {
         const toolbar = floatingToolbar(editorView.state, intl, {
           allowResizing: true,
           allowAnnotation: true,
-          appearance: 'full-page',
+          allowAdvancedToolBarOptions: true,
         });
 
-        const annotateToolbarComponent = toolbar!.items.find(
-          item => item.type === 'custom',
-        ) as FloatingToolbarCustom;
+        const annotateToolbarComponent = getToolbarItems(
+          toolbar!,
+          editorView,
+        ).find(item => item.type === 'custom') as FloatingToolbarCustom;
 
         const annotationToolbar = shallow(annotateToolbarComponent.render(
           editorView,