+++ packages/editor/editor-core/src/__tests__/visual-regression/common/panel.ts
@@ -1,38 +1,19 @@
-import {
-  initFullPageEditorWithAdf,
-  initCommentEditorWithAdf,
-  snapshot,
-  Device,
-} from '../_utils';
+import { initFullPageEditorWithAdf, snapshot } from '../_utils';
 import * as panel from './__fixtures__/panel-adf.json';
+import { Page } from '../../__helpers/page-objects/_types';
+import { panelSelectors } from '../../__helpers/page-objects/_panel';
 
-// skipping since new snapshot generated has problems
-// Expected image to be the same size as the snapshot (1280x436), but was different (1280x435).
-describe.skip('Panel overflow:', () => {
-  let page: any;
+describe('Panel:', () => {
+  let page: Page;
 
   beforeAll(() => {
     // @ts-ignore
     page = global.page;
   });
 
-  afterEach(async () => {
+  it('looks correct', async () => {
+    await initFullPageEditorWithAdf(page, panel);
+    await page.click(panelSelectors.panel);
     await snapshot(page);
-  });
-
-  it('looks correct for laptopMDPI', async () => {
-    await initFullPageEditorWithAdf(page, panel, Device.LaptopMDPI);
-  });
-
-  it('looks correct for ipad', async () => {
-    await initFullPageEditorWithAdf(page, panel, Device.iPad);
-  });
-
-  it('looks correct for iphone', async () => {
-    await initFullPageEditorWithAdf(page, panel, Device.iPhonePlus);
-  });
-
-  it('looks correct for LaptopMDPI in comment mode', async () => {
-    await initCommentEditorWithAdf(page, panel, Device.LaptopMDPI);
   });
 });