+++ packages/editor/editor-core/src/plugins/media/picker-facade.ts
@@ -19,10 +19,16 @@ import {
   MobileUploadEndEventPayload,
 } from './types';
 
-export type PickerType = 'popup' | 'customMediaPicker';
+export type PickerType =
+  | 'popup'
+  | 'clipboard'
+  | 'dropzone'
+  | 'customMediaPicker';
 export type ExtendedComponentConfigs = {
   popup: PopupConfig;
   customMediaPicker: CustomMediaPicker;
+  dropzone: null;
+  clipboard: null;
 };
 
 export type PickerFacadeConfig = {
@@ -39,6 +45,7 @@ export type MediaStateEventSubscriber = ((
 export type NewMediaEvent = (
   state: MediaState,
   onStateChanged: MediaStateEventSubscriber,
+  pickerType?: string,
 ) => void;
 
 export default class PickerFacade {
@@ -51,15 +58,18 @@ export default class PickerFacade {
     string,
     Array<MediaStateEventListener> | undefined
   > = {};
+  private analyticsName: string | undefined;
 
   constructor(
     pickerType: PickerType,
     readonly config: PickerFacadeConfig,
     readonly pickerConfig?: ExtendedComponentConfigs[PickerType],
     readonly mediaPickerFactoryClass = MediaPicker,
+    analyticsName?: string,
   ) {
     this.pickerType = pickerType;
     this.errorReporter = config.errorReporter;
+    this.analyticsName = analyticsName;
   }
 
   async init(): Promise<PickerFacade> {
@@ -75,6 +85,8 @@ export default class PickerFacade {
 
     (picker as any).on('upload-preview-update', this.handleUploadPreviewUpdate);
     (picker as any).on('upload-processing', this.handleReady);
+    // media picker not always fires upload-processing but always fires upload-end, and since handleReady() is idempotent it can be treated the same
+    (picker as any).on('upload-end', this.handleReady);
     (picker as any).on('upload-error', this.handleUploadError);
     (picker as any).on('mobile-upload-end', this.handleMobileUploadEnd);
 
@@ -98,6 +110,7 @@ export default class PickerFacade {
 
     (picker as any).removeAllListeners('upload-preview-update');
     (picker as any).removeAllListeners('upload-processing');
+    (picker as any).removeAllListeners('upload-end');
     (picker as any).removeAllListeners('upload-error');
 
     this.onStartListeners = [];
@@ -173,7 +186,11 @@ export default class PickerFacade {
 
     this.eventListeners[file.id] = [];
     this.onStartListeners.forEach(cb =>
-      cb(state, evt => this.subscribeStateChanged(file, evt)),
+      cb(
+        state,
+        evt => this.subscribeStateChanged(file, evt),
+        this.analyticsName || this.pickerType,
+      ),
     );
   };