+++ packages/elements/util-data-test/src/mention/mention-data.ts
@@ -2,7 +2,7 @@ declare var require: {
   <T>(path: string): T;
 };
 
-export const mentionData = require('../../json-data/mention-data.json') as any; // MentionsResult
+export const mentionData = require('../json-data/mention-data.json') as any; // MentionsResult
 
 export const mentionResult: any[] = mentionData.mentions;