+++ packages/search/global-search/src/__tests__/unit/common/ContainerSearchResults.test.tsx
@@ -356,7 +356,7 @@ const getConfluencePostQueryResults: (
     items: [],
     key: 'objects',
     title: messages.confluence_confluence_objects_heading,
-    showTotalSize: false,
+    showTotalSize: true,
     totalSize: 0,
   },
   {