+++ packages/core/avatar-group/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/avatar-group
 
+## 4.0.11
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
 ## 4.0.10
 
 ### Patch Changes