+++ packages/core/form/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/form
 
+## 6.1.11
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
 ## 6.1.10
 
 ### Patch Changes