+++ packages/core/global-navigation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/global-navigation
 
+## 7.6.2
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
 ## 7.6.1
 
 ### Patch Changes