+++ packages/core/icon-file-type/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/icon-file-type
 
+## 4.0.8
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
 ## 4.0.7
 
 ### Patch Changes