+++ packages/core/inline-message/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/inline-message
 
+## 10.0.13
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
 ## 10.0.12
 
 ### Patch Changes