+++ packages/core/multi-select/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/multi-select
 
+## 13.0.14
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
 ## 13.0.13
 
 ### Patch Changes