+++ packages/core/select/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/select
 
+## 10.0.7
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
 ## 10.0.6
 
 ### Patch Changes