+++ packages/core/textarea/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/textarea
 
+## 2.1.9
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
 ## 2.1.8
 
 ### Patch Changes