+++ packages/editor/editor-test-helpers/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @atlaskit/editor-test-helpers
 
+## 9.11.9
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
+## 9.11.8
+
+### Patch Changes
+
+- [patch][a82d6088e2](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a82d6088e2):
+
+  ED-4807 Use right cell type when spliting merged header cells
+
 ## 9.11.7
 
 ### Patch Changes