+++ packages/help/help/CHANGELOG.md
@@ -1,5 +1,21 @@
 ## 0.1.0
 
+## 0.14.3
+
+### Patch Changes
+
+- [patch][9e23c34ccb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9e23c34ccb):
+
+  Use generic types instead of type annotations for styles-components
+
+## 0.14.2
+
+### Patch Changes
+
+- [patch][ecca4d1dbb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ecca4d1dbb):
+
+  Upgraded Typescript to 3.3.x
+
 ## 0.14.1
 
 ### Patch Changes